Show dashboards for several projects
