+++ pages/api/metrics/[project_id].js
@@ -3,8 +3,9 @@ import { PSDB } from 'planetscale-node'
 
 const conn = new PSDB('main')
 
-export default async (_, res) => {
-  const [measures] = await conn.query('select * from metric')
+export default async (req, res) => {
+  const { project_id } = req.query
+  const [measures] = await conn.query('select * from metric where project_id = ?', project_id)
   res.statusCode = 200
   res.json(measures)
 }
